requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-dotenv==1.1.1
pdfplumber==0.11.7
google-genai==1.32.0
python-multipart==0.0.20
